+++ swap/src/bin/asb.rs
@@ -31,7 +31,8 @@ use swap::asb::config::{
     initial_setup, query_user_for_initial_config, read_config, Config, ConfigNotInitialized,
 };
 use swap::asb::{cancel, punish, redeem, refund, safely_abort, EventLoop, Finality, KrakenRate};
-use swap::common::check_latest_version;
+use swap::common::tracing_util::Format;
+use swap::common::{self, check_latest_version, get_logs};
 use swap::database::{open_db, AccessMode};
 use swap::network::rendezvous::XmrBtcNamespace;
 use swap::network::swarm;
@@ -40,44 +41,40 @@ use swap::protocol::alice::{run, AliceState};
 use swap::protocol::State;
 use swap::seed::Seed;
 use swap::tor::AuthenticatedClient;
-use swap::{asb, bitcoin, kraken, monero, tor};
+use swap::{bitcoin, kraken, monero, tor};
 use tracing_subscriber::filter::LevelFilter;
 
 const DEFAULT_WALLET_NAME: &str = "asb-wallet";
 
 #[tokio::main]
 async fn main() -> Result<()> {
+    // parse cli arguments
     let Arguments {
         testnet,
         json,
-        disable_timestamp,
         config_path,
         env_config,
         cmd,
     } = match parse_args(env::args_os()) {
         Ok(args) => args,
         Err(e) => {
+            // make sure to display the clap error message it exists
             if let Some(clap_err) = e.downcast_ref::<clap::Error>() {
-                match clap_err.kind {
-                    ErrorKind::HelpDisplayed | ErrorKind::VersionDisplayed => {
-                        println!("{}", clap_err.message);
-                        std::process::exit(0);
-                    }
-                    _ => {
-                        bail!(e);
-                    }
+                if let ErrorKind::HelpDisplayed | ErrorKind::VersionDisplayed = clap_err.kind {
+                    println!("{}", clap_err.message);
+                    std::process::exit(0);
                 }
             }
             bail!(e);
         }
     };
 
+    // warn if we're not on the latest version
     if let Err(e) = check_latest_version(env!("CARGO_PKG_VERSION")).await {
         eprintln!("{}", e);
     }
 
-    asb::tracing::init(LevelFilter::DEBUG, json, !disable_timestamp).expect("initialize tracing");
-
+    // read config from the specified path
     let config = match read_config(config_path.clone())? {
         Ok(config) => config,
         Err(ConfigNotInitialized {}) => {
@@ -86,6 +83,12 @@ async fn main() -> Result<()> {
         }
     };
 
+    // initialize tracing
+    let format = if json { Format::Json } else { Format::Raw };
+    let log_dir = config.data.dir.join("logs");
+    common::tracing_util::init(LevelFilter::DEBUG, format, log_dir).expect("initialize tracing");
+
+    // check for conflicting env / config values
     if config.monero.network != env_config.monero_network {
         bail!(format!(
             "Expected monero network in config file to be {:?} but was {:?}",
@@ -112,6 +115,7 @@ async fn main() -> Result<()> {
             rendezvous_addrs.sort();
             rendezvous_addrs.dedup();
             let new_len = rendezvous_addrs.len();
+
             if new_len < prev_len {
                 tracing::warn!(
                     "`rendezvous_point` config has {} duplicate entries, they are being ignored.",
@@ -119,9 +123,12 @@ async fn main() -> Result<()> {
                 );
             }
 
+            // initialize monero wallet
             let monero_wallet = init_monero_wallet(&config, env_config).await?;
             let monero_address = monero_wallet.get_main_address();
             tracing::info!(%monero_address, "Monero wallet address");
+
+            // check monero balance
             let monero = monero_wallet.get_balance().await?;
             match (monero.balance, monero.unlocked_balance) {
                 (0, _) => {
@@ -144,6 +151,7 @@ async fn main() -> Result<()> {
                 }
             }
 
+            // init bitcoin wallet
             let bitcoin_wallet = init_bitcoin_wallet(&config, &seed, env_config).await?;
             let bitcoin_balance = bitcoin_wallet.balance().await?;
             tracing::info!(%bitcoin_balance, "Bitcoin wallet balance");
@@ -313,6 +321,19 @@ async fn main() -> Result<()> {
         Command::Config => {
             let config_json = serde_json::to_string_pretty(&config)?;
             println!("{}", config_json);
+        }
+        Command::Logs {
+            logs_dir,
+            swap_id,
+            redact,
+        } => {
+            let dir = logs_dir.unwrap_or(config.data.dir.join("logs"));
+
+            let log_messages = get_logs(dir, swap_id, redact).await?;
+
+            for msg in log_messages {
+                println!("{msg}");
+            }
         }
         Command::WithdrawBtc { amount, address } => {
             let bitcoin_wallet = init_bitcoin_wallet(&config, &seed, env_config).await?;