+++ swap/src/cli.rs
@@ -3,7 +3,6 @@ pub mod cancel_and_refund;
 pub mod command;
 mod event_loop;
 mod list_sellers;
-pub mod tracing;
 pub mod transport;
 
 pub use behaviour::{Behaviour, OutEvent};