+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- ASB + CLI: You can now use the `logs` command to retrieve logs stored in the past, redacting addresses and id's using `logs --redact`.
+- ASB: The `--disable-timestamp` flag has been removed
+
 ## [0.13.4] - 2024-07-25
 
 - ASB: The `history` command can now be used while the asb is running.